refactor(user-service): refactor user-service for modularity and add CORS middleware
- Introduce a new setup for the user-service to improve structure, modularity, and maintainability.
- Added CORS middleware to handle cross-origin requests and moved route setups to a dedicated module.
- Refactored `main.go` to delegate server start logic to a new `StartServer` function.
